+++ test/cljds_test.dart
@@ -1,8 +1,3 @@
-import 'package:cljds/cljds.dart';
 import 'package:test/test.dart';
 
-void main() {
-  test('calculate', () {
-    expect(calculate(), 42);
-  });
-}
+void main() {}